+++ CHANGES.md
@@ -68,7 +68,7 @@
 
 * Multiple RunInference instances can now share the same model instance by setting the model_identifier parameter (Python) ([#31665](https://github.com/apache/beam/issues/31665)).
 * Removed a 3rd party LGPL dependency from the Go SDK ([#31765](https://github.com/apache/beam/issues/31765)).
-* Support for MapState and SetState when using Dataflow Runner with Streaming Engine (Java) ([[#18200](https://github.com/apache/beam/issues/18200)])
+* Support for MapState and SetState when using Dataflow Runner v1 with Streaming Engine (Java) ([[#18200](https://github.com/apache/beam/issues/18200)])
 
 ## Breaking Changes